chapters refactor

+++ mediaserviceinterfaces/src/main/java/com/liskovsoft/mediaserviceinterfaces/data/ChapterItem.java
@@ -2,6 +2,6 @@ package com.liskovsoft.mediaserviceinterfaces.data;
 
 public interface ChapterItem {
     String getTitle();
-    int getStartTimeMs();
+    long getStartTimeMs();
     String getCardImageUrl();
 }
